msg/async: move crypto handling from ProtocolV2 into AuthStreamHandler.
