+++ src/auth/cephx/CephxSessionHandler.cc
@@ -180,25 +180,3 @@ int CephxSessionHandler::check_message_signature(Message *m)
 
   return 0;
 }
-
-int CephxSessionHandler::encrypt_bufferlist(bufferlist &in, bufferlist &out) {
-  std::string error;
-  try {
-    key.encrypt(cct, in, out, &error);
-  } catch (std::exception &e) {
-    lderr(cct) << __func__ << " failed to encrypt buffer: " << error << dendl;
-    return -1;
-  }
-  return 0;
-}
-
-int CephxSessionHandler::decrypt_bufferlist(bufferlist &in, bufferlist &out) {
-  std::string error;
-  try {
-    key.decrypt(cct, in, out, &error);
-  } catch (std::exception &e) {
-    lderr(cct) << __func__ << " failed to decrypt buffer: " << error << dendl;
-    return -1;
-  }
-  return 0;
-}
