+++ src/auth/cephx/CephxSessionHandler.h
@@ -40,8 +40,5 @@ public:
 
   int sign_message(Message *m) override;
   int check_message_signature(Message *m) override ;
-
-  int encrypt_bufferlist(bufferlist &in, bufferlist &out);
-  int decrypt_bufferlist(bufferlist &in, bufferlist &out);
 };
